Rename process-cli -> backend-cli and embedded -> backend-embedded
