+++ backend-cli/src/main/assembly/dist-linux_aarch64.xml
@@ -6,10 +6,10 @@
         - the launcher script
           -->
 <assembly>
-  <id>windows</id>
+  <id>linux_aarch64</id>
   <includeBaseDirectory>false</includeBaseDirectory>
   <formats>
-    <format>zip</format>
+    <format>tar.gz</format>
   </formats>
 
   <!-- The dependencies -->
@@ -30,7 +30,7 @@
       <outputDirectory>lib</outputDirectory>
       <fileMode>0644</fileMode>
       <includes>
-        <include>**/sonarlint-process-cli-${project.version}.jar</include>
+        <include>**/${project.artifactId}-${project.version}.jar</include>
       </includes>
     </fileSet>
 
@@ -39,28 +39,44 @@
       <directory>src/main/assembly/bin</directory>
       <outputDirectory>bin</outputDirectory>
       <fileMode>0755</fileMode>
-      <lineEnding>dos</lineEnding>
+      <lineEnding>unix</lineEnding>
       <filtered>true</filtered>
       <includes>
-        <include>sonarlint-backend.bat</include>
+        <include>sonarlint-backend</include>
       </includes>
     </fileSet>
 
     <!-- jre basic, except bin/ -->
     <fileSet>
-      <directory>${unpack.dir}/windows/${jre.dirname.windows}</directory>
+      <directory>${unpack.dir}/linux_aarch64/${jre.dirname.linux}</directory>
       <outputDirectory>jre</outputDirectory>
       <excludes>
         <exclude>bin/**</exclude>
         <exclude>man/**</exclude>
+        <exclude>lib/jspawnhelper</exclude>
+        <exclude>lib/jexec</exclude>
         <exclude>plugin/**</exclude>
       </excludes>
     </fileSet>
 
-    <!-- jre bin -->
+    <!-- jre/bin/java -->
     <fileSet>
-      <directory>${unpack.dir}/windows/${jre.dirname.windows}/bin</directory>
+      <directory>${unpack.dir}/linux_aarch64/${jre.dirname.linux}/bin</directory>
       <outputDirectory>jre/bin</outputDirectory>
+      <includes>
+        <include>java</include>
+      </includes>
+      <fileMode>0755</fileMode>
+    </fileSet>
+
+    <!-- jre lib executable files -->
+    <fileSet>
+      <directory>${unpack.dir}/linux_aarch64/${jre.dirname.linux}/lib</directory>
+      <outputDirectory>jre/lib</outputDirectory>
+      <includes>
+        <include>jspawnhelper</include>
+        <include>jexec</include>
+      </includes>
       <fileMode>0755</fileMode>
     </fileSet>
   </fileSets>
