+++ backend-cli/src/main/assembly/dist-linux_x64.xml
@@ -6,7 +6,7 @@
         - the launcher script
           -->
 <assembly>
-  <id>linux_aarch64</id>
+  <id>linux_x64</id>
   <includeBaseDirectory>false</includeBaseDirectory>
   <formats>
     <format>tar.gz</format>
@@ -30,7 +30,7 @@
       <outputDirectory>lib</outputDirectory>
       <fileMode>0644</fileMode>
       <includes>
-        <include>**/sonarlint-process-cli-${project.version}.jar</include>
+        <include>**/${project.artifactId}-${project.version}.jar</include>
       </includes>
     </fileSet>
 
@@ -48,7 +48,7 @@
 
     <!-- jre basic, except bin/ -->
     <fileSet>
-      <directory>${unpack.dir}/linux_aarch64/${jre.dirname.linux}</directory>
+      <directory>${unpack.dir}/linux_x64/${jre.dirname.linux}</directory>
       <outputDirectory>jre</outputDirectory>
       <excludes>
         <exclude>bin/**</exclude>
@@ -61,7 +61,7 @@
 
     <!-- jre/bin/java -->
     <fileSet>
-      <directory>${unpack.dir}/linux_aarch64/${jre.dirname.linux}/bin</directory>
+      <directory>${unpack.dir}/linux_x64/${jre.dirname.linux}/bin</directory>
       <outputDirectory>jre/bin</outputDirectory>
       <includes>
         <include>java</include>
@@ -71,7 +71,7 @@
 
     <!-- jre lib executable files -->
     <fileSet>
-      <directory>${unpack.dir}/linux_aarch64/${jre.dirname.linux}/lib</directory>
+      <directory>${unpack.dir}/linux_x64/${jre.dirname.linux}/lib</directory>
       <outputDirectory>jre/lib</outputDirectory>
       <includes>
         <include>jspawnhelper</include>
