+++ backend-cli/src/main/assembly/dist-macosx-x86_64.xml
@@ -6,7 +6,7 @@
         - the launcher script
           -->
 <assembly>
-  <id>linux_x64</id>
+  <id>macosx_x64</id>
   <includeBaseDirectory>false</includeBaseDirectory>
   <formats>
     <format>tar.gz</format>
@@ -30,7 +30,7 @@
       <outputDirectory>lib</outputDirectory>
       <fileMode>0644</fileMode>
       <includes>
-        <include>**/sonarlint-process-cli-${project.version}.jar</include>
+        <include>**/${project.artifactId}-${project.version}.jar</include>
       </includes>
     </fileSet>
 
@@ -46,22 +46,20 @@
       </includes>
     </fileSet>
 
-    <!-- jre basic, except bin/ -->
+    <!-- jre basic, except bin/ and misc -->
     <fileSet>
-      <directory>${unpack.dir}/linux_x64/${jre.dirname.linux}</directory>
+      <directory>${unpack.dir}/macosx_x64/${jre.dirname.macosx}</directory>
       <outputDirectory>jre</outputDirectory>
       <excludes>
         <exclude>bin/**</exclude>
         <exclude>man/**</exclude>
         <exclude>lib/jspawnhelper</exclude>
-        <exclude>lib/jexec</exclude>
-        <exclude>plugin/**</exclude>
       </excludes>
     </fileSet>
 
-    <!-- jre/bin/java -->
+    <!-- jre bin -->
     <fileSet>
-      <directory>${unpack.dir}/linux_x64/${jre.dirname.linux}/bin</directory>
+      <directory>${unpack.dir}/macosx_x64/${jre.dirname.macosx}/bin</directory>
       <outputDirectory>jre/bin</outputDirectory>
       <includes>
         <include>java</include>
@@ -71,11 +69,10 @@
 
     <!-- jre lib executable files -->
     <fileSet>
-      <directory>${unpack.dir}/linux_x64/${jre.dirname.linux}/lib</directory>
+      <directory>${unpack.dir}/macosx_x64/${jre.dirname.macosx}/lib</directory>
       <outputDirectory>jre/lib</outputDirectory>
       <includes>
         <include>jspawnhelper</include>
-        <include>jexec</include>
       </includes>
       <fileMode>0755</fileMode>
     </fileSet>
