+++ osgi/pom.xml
@@ -17,12 +17,12 @@
   <dependencies>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>sonarlint-embedded</artifactId>
+      <artifactId>sonarlint-backend-embedded</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>sonarlint-embedded</artifactId>
+      <artifactId>sonarlint-backend-embedded</artifactId>
       <version>${project.version}</version>
       <classifier>sources</classifier>
     </dependency>
